requirements.txt
fastapi
uvicorn
google-generativeai
mistralai
diffusers
transformers
pillow
python-dotenv
